# crates/ryu-sdk/specs/gateway-openapi.yaml
openapi: 3.1.0
info:
  title: Ryu Gateway API
  version: "1.0.0"
  description: |
    The **Ryu Gateway** is the OpenAI-compatible control layer that governs every
    model call: routing, fallback, rate-limiting, firewall/PII-DLP, budgets, evals,
    and audit. Point any OpenAI-compatible client at the gateway by swapping its
    base URL; the request surface below is a superset of the OpenAI API plus Ryu's
    governance, eval, and audit endpoints.

    Authentication is via a bearer token (`Authorization: Bearer <api-key>`). When
    the gateway runs with `auth.require_auth = false`, anonymous requests are
    allowed. Admin endpoints (`/v1/config`, `/v1/audit`) additionally require the
    master key or a loopback peer.
servers:
  - url: http://localhost:7981
    description: Local gateway (default bind)
security:
  - bearerAuth: []
...
paths:
...
  /v1/audio/transcriptions:
    post:
      tags: [Inference]
      summary: Transcribe audio (speech-to-text)
      operationId: createTranscription
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/TranscriptionRequest"
      responses:
        "200":
          description: Transcribed text.
          content:
            application/json:
              schema:
                type: object
                properties:
                  text: { type: string }
                required: [text]
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key issued by the gateway, or the master key for admin endpoints.
...
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            type: { type: string }
            code: { type: string }
...
    TranscriptionRequest:
      type: object
      properties:
        file: { type: string, format: binary }
        model: { type: string }
        language: { type: string }
        prompt: { type: string }
        response_format: { type: string, enum: [json, text, vtt, srt] }
        temperature: { type: number, format: float }
      required: [file, model]
...